requirements/base.txt
# Django
Django>=5.0,<5.1

# Django REST Framework
djangorestframework>=3.14,<4.0

# JWT Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment Variables
django-environ>=0.11,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# CORS Headers
django-cors-headers>=4.3,<5.0

# Password Hashing (bcrypt)
bcrypt>=4.1,<5.0

# API Schema/Documentation
drf-spectacular>=0.27,<1.0

# UUID support
uuid>=1.30

